feat: Add support for configuring Terragrunt log options
This commit makes the following key changes:

- Adds support for configuring Terragrunt log options, including setting the log level and forwarding Terraform stdout.
- Enables the following environment variables in the `TestTerragruntExecWithPlanOutput` test:
  - `TF_VAR_project_name`, `TF_VAR_environment`, `TF_VAR_region`, `TF_VAR_instance_type`, `TF_VAR_db_name`
  - `GOOGLE_CLOUD_KEYFILE_JSON`, `AWS_ACCESS_KEY_ID`, `AWS_SECRET_ACCESS_KEY`, `DOCKER_REGISTRY`
- Introduces a new interface `Cmd` in the `iac_cmd.go` file, which defines the common methods for executing commands in Infrastructure as Code (IaC) tools like Terraform, Terragrunt, and OpenTofu within a Dagger container.

The purpose of these changes is to enhance the Terragrunt test suite by providing more control over the logging and environment variables used during the tests, improving the overall testability and debuggability of the Terragrunt module.
